perf(QueryBuilder): Use forward references to fix query builder circular import
The first attempt to split the imports under the various if conditions didn't
work. Instead it made everything slower (I'm not sure why this is the case).
This is inspired by https://peps.python.org/pep-0484/#forward-references. This
allows us lift the import to the top so it is done at start up instead of during
the query.
