+++ mypy.ini
@@ -97,7 +97,6 @@ files = fixtures/mypy-stubs,
         src/sentry/search/base.py,
         src/sentry/search/events/builder.py,
         src/sentry/search/events/constants.py,
-        src/sentry/search/events/datasets/profiles.py,
         src/sentry/search/events/types.py,
         src/sentry/search/snuba/,
         src/sentry/sentry_metrics/,
