+++ src/sentry/search/events/builder.py
@@ -9,6 +9,7 @@ from typing import (
     Mapping,
     Match,
     Optional,
+    Protocol,
     Sequence,
     Set,
     Tuple,
@@ -70,6 +71,11 @@ from sentry.search.events.constants import (
     VALID_FIELD_PATTERN,
 )
 from sentry.search.events.datasets.base import DatasetConfig
+from sentry.search.events.datasets.discover import DiscoverDatasetConfig
+from sentry.search.events.datasets.metrics import MetricsDatasetConfig
+from sentry.search.events.datasets.metrics_layer import MetricsLayerDatasetConfig
+from sentry.search.events.datasets.profiles import ProfilesDatasetConfig
+from sentry.search.events.datasets.sessions import SessionsDatasetConfig
 from sentry.search.events.fields import (
     ColumnArg,
     FunctionDetails,
@@ -271,25 +277,15 @@ class QueryBuilder:
     ]:
         self.config: DatasetConfig
         if self.dataset in [Dataset.Discover, Dataset.Transactions, Dataset.Events]:
-            from sentry.search.events.datasets.discover import DiscoverDatasetConfig
-
             self.config = DiscoverDatasetConfig(self)
         elif self.dataset == Dataset.Sessions:
-            from sentry.search.events.datasets.sessions import SessionsDatasetConfig
-
             self.config = SessionsDatasetConfig(self)
         elif self.dataset in [Dataset.Metrics, Dataset.PerformanceMetrics]:
             if self.use_metrics_layer:
-                from sentry.search.events.datasets.metrics_layer import MetricsLayerDatasetConfig
-
                 self.config = MetricsLayerDatasetConfig(self)
             else:
-                from sentry.search.events.datasets.metrics import MetricsDatasetConfig
-
                 self.config = MetricsDatasetConfig(self)
         elif self.dataset == Dataset.Profiles:
-            from sentry.search.events.datasets.profiles import ProfilesDatasetConfig
-
             self.config = ProfilesDatasetConfig(self)
         else:
             raise NotImplementedError(f"Data Set configuration not found for {self.dataset}.")
@@ -2802,3 +2798,56 @@ class TimeseriesMetricQueryBuilder(MetricsQueryBuilder):
             "data": list(time_map.values()),
             "meta": [{"name": key, "type": value} for key, value in meta_dict.items()],
         }
+
+
+class ProfilesQueryBuilderProtocol(Protocol):
+    @property
+    def config(self) -> ProfilesDatasetConfig:
+        ...
+
+    @property
+    def params(self) -> ParamsType:
+        ...
+
+    def column(self, name: str) -> Column:
+        ...
+
+    def resolve_params(self) -> List[WhereType]:
+        ...
+
+
+class ProfilesQueryBuilderMixin:
+    def resolve_column_name(self: ProfilesQueryBuilderProtocol, col: str) -> str:
+        # giving resolved a type here convinces mypy that the type is str
+        resolved: str = self.config.resolve_column(col)
+        return resolved
+
+    def resolve_params(self: ProfilesQueryBuilderProtocol) -> List[WhereType]:
+        # not sure how to make mypy happy here as `super()`
+        # refers to the other parent query builder class
+        conditions: List[WhereType] = super().resolve_params()  # type: ignore
+
+        # the profiles dataset requires a condition
+        # on the organization_id in the query
+        conditions.append(
+            Condition(
+                self.column("organization.id"),
+                Op.EQ,
+                self.params["organization_id"],
+            )
+        )
+
+        return conditions
+
+    def get_field_type(self: ProfilesQueryBuilderProtocol, field: str) -> Optional[str]:
+        # giving resolved a type here convinces mypy that the type is str
+        resolved: Optional[str] = self.config.resolve_column_type(field)
+        return resolved
+
+
+class ProfilesQueryBuilder(ProfilesQueryBuilderMixin, QueryBuilder):
+    pass
+
+
+class ProfilesTimeseriesQueryBuilder(ProfilesQueryBuilderMixin, TimeseriesQueryBuilder):
+    pass
