+++ src/sentry/search/events/datasets/field_aliases.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from typing import Any, List, Tuple
 
 import sentry_sdk
@@ -6,13 +8,14 @@ from snuba_sdk import AliasedExpression, Function
 from sentry.discover.models import TeamKeyTransaction
 from sentry.exceptions import IncompatibleMetricsQuery
 from sentry.models import ProjectTeam
-from sentry.search.events import constants, fields
-from sentry.search.events.builder import QueryBuilder
+from sentry.search.events import builder, constants, fields
 from sentry.search.events.types import SelectType
 from sentry.utils.numbers import format_grouped_length
 
 
-def dry_run_default(builder: QueryBuilder, alias: str, *args: Any, **kwargs: Any) -> SelectType:
+def dry_run_default(
+    builder: builder.QueryBuilder, alias: str, *args: Any, **kwargs: Any
+) -> SelectType:
     """It doesn't really matter what we return here, the query won't be run
 
     This is so we can easily swap to something when we're dry running to prevent hitting postgres at all
@@ -21,7 +24,7 @@ def dry_run_default(builder: QueryBuilder, alias: str, *args: Any, **kwargs: Any
 
 
 def resolve_team_key_transaction_alias(
-    builder: QueryBuilder, resolve_metric_index: bool = False
+    builder: builder.QueryBuilder, resolve_metric_index: bool = False
 ) -> SelectType:
     team_key_transactions = get_team_transactions(builder, resolve_metric_index)
     if len(team_key_transactions) == 0:
@@ -38,7 +41,7 @@ def resolve_team_key_transaction_alias(
 
 
 def get_team_transactions(
-    builder: QueryBuilder, resolve_metric_index: bool = False
+    builder: builder.QueryBuilder, resolve_metric_index: bool = False
 ) -> List[Tuple[int, str]]:
     org_id = builder.params.get("organization_id")
     project_ids = builder.params.get("project_id")
@@ -84,7 +87,7 @@ def get_team_transactions(
     return team_key_transactions
 
 
-def resolve_project_slug_alias(builder: QueryBuilder, alias: str) -> SelectType:
+def resolve_project_slug_alias(builder: builder.QueryBuilder, alias: str) -> SelectType:
     builder.value_resolver_map[alias] = lambda project_id: builder.project_ids.get(project_id, "")
     builder.meta_resolver_map[alias] = "string"
     return AliasedExpression(exp=builder.column("project_id"), alias=alias)
