+++ src/sentry/search/events/datasets/filter_aliases.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from functools import reduce
 from typing import List, Mapping, Optional
 
@@ -6,8 +8,7 @@ from snuba_sdk import Condition, Op
 from sentry.api.event_search import SearchFilter, SearchKey, SearchValue
 from sentry.exceptions import InvalidSearchQuery
 from sentry.models import Environment, Release, SemverFilter
-from sentry.search.events import constants
-from sentry.search.events.builder import QueryBuilder
+from sentry.search.events import builder, constants
 from sentry.search.events.filter import (
     _flip_field_sort,
     handle_operator_negation,
@@ -19,7 +20,9 @@ from sentry.search.utils import parse_release
 from sentry.utils.strings import oxfordize_list
 
 
-def team_key_transaction_filter(builder: QueryBuilder, search_filter: SearchFilter) -> WhereType:
+def team_key_transaction_filter(
+    builder: builder.QueryBuilder, search_filter: SearchFilter
+) -> WhereType:
     value = search_filter.value.value
     key_transaction_expr = builder.resolve_field_alias(constants.TEAM_KEY_TRANSACTION_ALIAS)
 
@@ -37,7 +40,9 @@ def team_key_transaction_filter(builder: QueryBuilder, search_filter: SearchFilt
     )
 
 
-def release_filter_converter(builder: QueryBuilder, search_filter: SearchFilter) -> WhereType:
+def release_filter_converter(
+    builder: builder.QueryBuilder, search_filter: SearchFilter
+) -> WhereType:
     """Parse releases for potential aliases like `latest`"""
     if search_filter.value.is_wildcard():
         operator = search_filter.operator
@@ -65,7 +70,7 @@ def release_filter_converter(builder: QueryBuilder, search_filter: SearchFilter)
 
 
 def project_slug_converter(
-    builder: QueryBuilder, search_filter: SearchFilter
+    builder: builder.QueryBuilder, search_filter: SearchFilter
 ) -> Optional[WhereType]:
     """Convert project slugs to ids and create a filter based on those.
     This is cause we only store project ids in clickhouse.
@@ -106,7 +111,7 @@ def project_slug_converter(
 
 
 def release_stage_filter_converter(
-    builder: QueryBuilder, search_filter: SearchFilter
+    builder: builder.QueryBuilder, search_filter: SearchFilter
 ) -> Optional[WhereType]:
     """
     Parses a release stage search and returns a snuba condition to filter to the
@@ -142,7 +147,7 @@ def release_stage_filter_converter(
 
 
 def semver_filter_converter(
-    builder: QueryBuilder, search_filter: SearchFilter
+    builder: builder.QueryBuilder, search_filter: SearchFilter
 ) -> Optional[WhereType]:
     """
     Parses a semver query search and returns a snuba condition to filter to the
@@ -214,7 +219,7 @@ def semver_filter_converter(
 
 
 def semver_package_filter_converter(
-    builder: QueryBuilder, search_filter: SearchFilter
+    builder: builder.QueryBuilder, search_filter: SearchFilter
 ) -> Optional[WhereType]:
     """
     Applies a semver package filter to the search. Note that if the query returns more than
@@ -243,7 +248,7 @@ def semver_package_filter_converter(
 
 
 def semver_build_filter_converter(
-    builder: QueryBuilder, search_filter: SearchFilter
+    builder: builder.QueryBuilder, search_filter: SearchFilter
 ) -> Optional[WhereType]:
     """
     Applies a semver build filter to the search. Note that if the query returns more than
