+++ src/sentry/snuba/profiles.py
@@ -1,13 +1,9 @@
 from datetime import datetime
-from typing import Any, List, Optional, Protocol
+from typing import Any, List, Optional
 
-from snuba_sdk.column import Column
-from snuba_sdk.conditions import Condition, Op
-
-from sentry.search.events.builder import QueryBuilder, TimeseriesQueryBuilder
-from sentry.search.events.datasets.profiles import ProfilesDatasetConfig
+from sentry.search.events.builder import ProfilesQueryBuilder, ProfilesTimeseriesQueryBuilder
 from sentry.search.events.fields import InvalidSearchQuery, get_json_meta_type
-from sentry.search.events.types import ParamsType, WhereType
+from sentry.search.events.types import ParamsType
 from sentry.snuba.discover import transform_tips, zerofill
 from sentry.utils.snuba import Dataset, SnubaTSResult
 
@@ -94,56 +90,3 @@ def timeseries_query(
         params["end"],
         rollup,
     )
-
-
-class ProfilesQueryBuilderProtocol(Protocol):
-    @property
-    def config(self) -> ProfilesDatasetConfig:
-        ...
-
-    @property
-    def params(self) -> ParamsType:
-        ...
-
-    def column(self, name: str) -> Column:
-        ...
-
-    def resolve_params(self) -> List[WhereType]:
-        ...
-
-
-class ProfilesQueryBuilderMixin:
-    def resolve_column_name(self: ProfilesQueryBuilderProtocol, col: str) -> str:
-        # giving resolved a type here convinces mypy that the type is str
-        resolved: str = self.config.resolve_column(col)
-        return resolved
-
-    def resolve_params(self: ProfilesQueryBuilderProtocol) -> List[WhereType]:
-        # not sure how to make mypy happy here as `super()`
-        # refers to the other parent query builder class
-        conditions: List[WhereType] = super().resolve_params()  # type: ignore
-
-        # the profiles dataset requires a condition
-        # on the organization_id in the query
-        conditions.append(
-            Condition(
-                self.column("organization.id"),
-                Op.EQ,
-                self.params["organization_id"],
-            )
-        )
-
-        return conditions
-
-    def get_field_type(self: ProfilesQueryBuilderProtocol, field: str) -> Optional[str]:
-        # giving resolved a type here convinces mypy that the type is str
-        resolved: Optional[str] = self.config.resolve_column_type(field)
-        return resolved
-
-
-class ProfilesQueryBuilder(ProfilesQueryBuilderMixin, QueryBuilder):
-    pass
-
-
-class ProfilesTimeseriesQueryBuilder(ProfilesQueryBuilderMixin, TimeseriesQueryBuilder):
-    pass
